requirements.txt
argcomplete==1.9.3
astroid==2.2.2
boto3==1.9.108
botocore==1.12.108
certifi==2018.11.29
cfn-flip==1.1.0.post1
chardet==3.0.4
Click==7.0
Django==2.1.7
django-grappelli==2.12.2
django-js-asset==1.2.2
djangorestframework==3.9.2
docutils==0.14
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.8
isort==4.3.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
placebo==0.9.0
pylint==2.3.1
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.9
PyYAML==3.13
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.5
typed-ast==1.3.1
Unidecode==1.0.23
urllib3==1.24.1
Werkzeug==0.14.1
whitenoise==4.1.2
wrapt==1.11.1
wsgi-request-logger==0.4.6
